wave_2/snicket/pyproject.toml
[tool.poetry]
name = "raid"
version = "0.1.0"
description = ""
authors = ["fabe1847 <fabian.beez@tu-ilmenau.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.67.1"
ipykernel = "^6.29.5"
torch = "^2.7.0"
torchvision = "^0.22.0"
pillow = "^11.2.1"
timm = "^1.0.15"
optuna = "^4.3.0"
opencv-python = "^4.11.0.86"
matplotlib = "^3.10.3"
grad-cam = "^1.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
